+++ R/plot.gcfit.R
@@ -2,9 +2,10 @@
 #'
 #' @param x A fit for some growth data
 #' @param y Not used
-#' @param show_raw Whether or not to show the original data (default \code{TRUE})
+#' @param show_data Whether or not to show the original data (default \code{TRUE})
 #' @param show_maxrate Whether or not to show a tangent line where the maximum growth rate occurs (default \code{TRUE})
 #' @param show_asymptote Whether or not to indicate the maximum growth level (default \code{FALSE})
+#' @param show_lag Whether or not to indicate where lag phase ends (default \code{FALSE}
 #' @param ... Optional arguments to plot. Note that \pkg{grofit} does not pass these along to the base graphics, so they're mostly useless at the moment.
 #'
 #' @export
@@ -15,33 +16,34 @@
 #' lfit <- fit_growth_logistic(df=mydata, Time, OD600)
 #' plot(lfit)}
 #' 
-plot.gcfit <- function(x, y=NULL, show_raw=TRUE, show_maxrate=TRUE,
-                       show_asymptote=FALSE, ...) {
+plot.gcfit <- function(x, y=NULL, show_data=TRUE, show_maxrate=TRUE,
+                       show_asymptote=FALSE, show_lag=FALSE, ...) {
     opt_args <- list(...)
 
     opt_args$xlab <- ifelse("xlab" %in% names(opt_args), opt_args$xlab,
-                            x$raw$time_col)
+                            x$data$time_col)
     opt_args$ylab <- ifelse("ylab" %in% names(opt_args), opt_args$ylab,
-                            x$raw$data_col)
+                            x$data$data_col)
     opt_args$x <- x$fit$time
     opt_args$y <- x$fit$data
     opt_args$type <- "l"
 
     opt_args$show_maxrate <- NULL
-    opt_args$show_raw <- NULL
+    opt_args$show_data <- NULL
     opt_args$show_asymptote <- NULL
 
     # Plot the fitted curve
     try(do.call(plot, opt_args))
 
     # Add the raw data points
-    if(show_raw) {
-        points(x$raw$df[[x$raw$time_col]], x$raw$df[[x$raw$data_col]], ...)
+    if(show_data) {
+        points(x$data$df[[x$data$time_col]], x$data$df[[x$data$data_col]], ...)
     }
 
-    try(abline(a = x$parameters$lag_length[[1]],
-               b = x$parameters$max_rate[[1]]))
-
+    if (show_lag) {
+        try(abline(v = x$parameters$lag_length[[1]]))
+    }
+    
     # Add a tangent line where the maximum growth rate occurs
     if(show_maxrate) {
         yvals <- (x$fit$time * x$parameters$max_rate[[1]]) + (-1 * x$parameters$max_rate[[1]] * x$parameters$lag_length[[1]])